# Fix network for tooltracking based on monai rc2 + optimize transform