--- a/sample-apps/endoscopy/lib/infers/tooltracking.py
+++ b/sample-apps/endoscopy/lib/infers/tooltracking.py
@@ -45,8 +45,10 @@ class ToolTracking(InferTask):
         labels=None,
         dimension=2,
         description="A pre-trained semantic segmentation model for Tool Tracking",
+        find_contours=True,
         **kwargs,
     ):
+        self.find_contours = find_contours
         super().__init__(
             path=path,
             network=network,
@@ -75,14 +77,20 @@ class ToolTracking(InferTask):
         return SimpleInferer()
 
     def post_transforms(self, data=None) -> Sequence[Callable]:
-        return [
+        t = [
             EnsureTyped(keys="pred", device=data.get("device") if data else None),
             AsDiscreted(keys="pred", argmax=True),
             Restored(keys="pred", ref_image="image"),
             SqueezeDimd(keys="pred", dim=0),
-            FindContoursd(keys="pred", labels=self.labels),
         ]
 
+        if self.find_contours:
+            t.append(FindContoursd(keys="pred", labels=self.labels))
+        return t
+
     def writer(self, data, extension=None, dtype=None):
-        writer = PolygonWriter(label=self.output_label_key, json=self.output_json_key)
-        return writer(data)
+        if self.find_contours:
+            writer = PolygonWriter(label=self.output_label_key, json=self.output_json_key)
+            return writer(data)
+
+        return super().writer(data, extension, dtype)