… sample-apps/endoscopy/lib/net/__init__.py
… sample-apps/endoscopy/lib/net/ranzcrnet.py