Signed-off-by: Sachidanand Alle <sachidanand.alle@gmail.com>

Input Image size: (1920, 1080, 3)
Input Model: Tool Tracking

- Total e2e Latency = 180 ms
- Pre Transform Latency = 60 ms (30 ms for Load; 20 ms for resize)
- Infer Latency = 40 ms
- Post Transform = 80 ms (80 ms for Find Contour)

Exclude Find Contour in Scoring Task (epistemic)... 15% speed up in overall time