@SachidanandAlle 2022-09-13
@finalelement the previous PR is not correctly verified: https://github.com/Project-MONAI/MONAILabel/pull/978

`VarianceMetric` from monai core...
It doesn't work and throws error for NAN values.. so did kind of rollback to use np.nanvar (based on function arg)

Check for options in torch.var to ignore NaN values